requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-tailwind==1.0.3
dj-database-url==2.2.0
Django==4.2.13
django-appconf==1.0.6
django-compressor==4.4
django-crispy-forms==2.1
django-filter==24.3
django-imagekit==5.0.0
django-taggit==6.1.0
django-unfold==0.23.0
gunicorn==22.0.0
idna==3.10
packaging==24.2
pilkit==3.0
pillow==10.3.0
pip3-autoremove==1.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rcssmin==1.1.1
requests==2.32.3
rjsmin==1.2.1
setuptools==75.8.2
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.6.0
